Update `getLogDownloadButton` params to match `fetchLogs`
Update the `getLogDownloadButton` function to receive the same
params we send to `fetchLogs`. This makes it more consistent and
also more useful for 3rd-party consumers who may want to conditionally
display the button(s) based on the step / TaskRun status.
